Add support for extProperty config key to expose model as an extra property under project.ext (issue #102)
