+++ src/main/groovy/com/gorylenko/properties/BuildTimeProperty.groovy
@@ -6,19 +6,17 @@ import java.time.Instant
 import org.ajoberstar.grgit.Grgit
 
 class BuildTimeProperty extends Closure<String> {
-    private Instant buildTime
     private String dateFormat
     private String timezone
 
-    BuildTimeProperty(Instant buildTime, String dateFormat, String timezone) {
+    BuildTimeProperty(String dateFormat, String timezone) {
         super(null)
-        this.buildTime = buildTime
         this.dateFormat = dateFormat
         this.timezone = timezone
     }
 
     String doCall(Grgit repo) {
-        return formatDate(buildTime, dateFormat, timezone)
+        return formatDate(Instant.now(), dateFormat, timezone)
     }
 
     private String formatDate(Instant instant, String dateFormat, String timezone) {
